move containsIgnoreCase from Ascii -> Strings
