+++ preferencesfx/src/main/java/com/dlsc/preferencesfx/util/Strings.java
@@ -58,4 +58,29 @@ public class Strings {
     return new String(hexChars);
   }
 
+  /**
+   * Indicates whether the character sequence {@code sequence} contains the {@code subSequence},
+   * ignoring the case in the same manner as {@link String#equalsIgnoreCase(String)}.
+   *
+   * @param str       the string to search in
+   * @param searchStr the string to search for
+   * @return {@code true} if str contains the searchStr (case insensitive)
+   */
+  static boolean containsIgnoreCase(String str, String searchStr) {
+    if (str == null || searchStr == null) {
+      return false;
+    }
+    if (searchStr.isEmpty()) {
+      return true;
+    }
+
+    final int len = searchStr.length();
+    final int max = str.length() - len;
+    for (int i = 0; i <= max; i++) {
+      if (str.regionMatches(true, i, searchStr, 0, len)) {
+        return true;
+      }
+    }
+    return false;
+  }
 }
